--- tests/director-tests/go.mod
@@ -1,47 +1,0 @@
-module github.com/kyma-incubator/compass/tests/director-tests
-
-go 1.13
-
-require (
-	github.com/99designs/gqlgen v0.10.2
-	github.com/Masterminds/goutils v1.1.0
-	github.com/Masterminds/semver v1.5.0 // indirect
-	github.com/Masterminds/sprig v2.22.0+incompatible
-	github.com/agnivade/levenshtein v1.0.3
-	github.com/asaskevich/govalidator v0.0.0-20180315120708-ccb8e960c48f
-	github.com/avast/retry-go v2.4.3+incompatible
-	github.com/davecgh/go-spew v1.1.1
-	github.com/dgrijalva/jwt-go v3.2.0+incompatible
-	github.com/go-ozzo/ozzo-validation v3.6.0+incompatible
-	github.com/gogo/protobuf v1.3.1
-	github.com/google/gofuzz v1.0.0
-	github.com/google/uuid v1.1.1
-	github.com/hashicorp/errwrap v1.0.0
-	github.com/hashicorp/go-multierror v1.0.0
-	github.com/huandu/xstrings v1.2.1
-	github.com/imdario/mergo v0.3.8
-	github.com/kisielk/errcheck v1.2.0
-	github.com/konsorten/go-windows-terminal-sequences v1.0.2
-	github.com/kyma-incubator/compass v0.0.0-20200103100931-6750d660a361
-	github.com/machinebox/graphql v0.2.3-0.20181106130121-3a9253180225
-	github.com/mitchellh/copystructure v1.0.0
-	github.com/mitchellh/reflectwalk v1.0.1
-	github.com/pkg/errors v0.8.1
-	github.com/pmezard/go-difflib v1.0.0
-	github.com/sirupsen/logrus v1.4.2
-	github.com/stretchr/testify v1.4.0
-	github.com/vektah/gqlparser v1.2.0
-	github.com/vrischmann/envconfig v1.2.0
-	github.com/xeipuuv/gojsonpointer v0.0.0-20190905194746-02993c407bfb
-	github.com/xeipuuv/gojsonreference v0.0.0-20180127040603-bd5ef7bd5415
-	github.com/xeipuuv/gojsonschema v1.2.0
-	golang.org/x/crypto v0.0.0-20191227163750-53104e6ec876
-	golang.org/x/net v0.0.0-20191209160850-c0dbc17a3553
-	golang.org/x/sys v0.0.0-20191228213918-04cbcbbfeed8
-	golang.org/x/text v0.3.2
-	golang.org/x/tools v0.0.0-20191227053925-7b8e75db28f4
-	gopkg.in/inf.v0 v0.9.1
-	gopkg.in/yaml.v2 v2.2.7
-	k8s.io/apimachinery v0.17.0
-	k8s.io/klog v1.0.0
-)
