+++ tests/director/api-tests/runtime_api_test.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/kyma-incubator/compass/tests/director-tests/pkg/ptr"
+	"github.com/kyma-incubator/compass/tests/director/pkg/ptr"
 
 	"github.com/google/uuid"
