+++ tests/director/api-tests/scopes_test.go
