+++ tests/director/gateway-integration-tests/compass_auth_test.go
@@ -1,4 +1,4 @@
-package auth_related_tests
+package gateway_integration_tests
 
 import (
 	"context"
@@ -10,11 +10,11 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"github.com/kyma-incubator/compass/tests/director-tests/pkg/gql"
+	"github.com/kyma-incubator/compass/tests/director/pkg/gql"
 
 	"github.com/kyma-incubator/compass/components/director/pkg/graphql"
 
-	"github.com/kyma-incubator/compass/tests/director-tests/pkg/idtokenprovider"
+	"github.com/kyma-incubator/compass/tests/director/pkg/idtokenprovider"
 	"github.com/stretchr/testify/require"
 )
