+++ tests/director/gateway-integration-tests/integration_system_scope_test.go
