+++ tests/director/gateway-integration-tests/playground_test.go
@@ -1,4 +1,4 @@
-package auth_related_tests
+package gateway_integration_tests
 
 import (
 	"context"
@@ -10,9 +10,9 @@ import (
 	"time"
 
 	"github.com/kyma-incubator/compass/components/director/pkg/graphql"
-	"github.com/kyma-incubator/compass/tests/director-tests/pkg/connector"
-	"github.com/kyma-incubator/compass/tests/director-tests/pkg/gql"
-	"github.com/kyma-incubator/compass/tests/director-tests/pkg/idtokenprovider"
+	"github.com/kyma-incubator/compass/tests/director/pkg/connector"
+	"github.com/kyma-incubator/compass/tests/director/pkg/gql"
+	"github.com/kyma-incubator/compass/tests/director/pkg/idtokenprovider"
 	gcli "github.com/machinebox/graphql"
 	"github.com/stretchr/testify/require"
 	"github.com/vrischmann/envconfig"
