+++ tests/director/gateway-integration-tests/token_utils.go
@@ -1,4 +1,4 @@
-package e2e
+package gateway_integration_tests
 
 import (
 	"crypto/tls"
