+++ tests/director/gen-examples.sh
