+++ tests/director/pkg/connector/queries.go
