+++ tests/director/pkg/ptr/ptr.go
