+++ tests/director/run.sh
@@ -17,4 +17,4 @@ GATEWAY_OAUTH20_SUBDOMAIN="compass-gateway-auth-oauth" \
 GATEWAY_JWT_SUBDOMAIN="compass-gateway" \
 GATEWAY_CLIENT_CERTS_SUBDOMAIN="compass-gateway-mtls" \
 DEFAULT_TENANT="3e64ebae-38b5-46a0-b1ed-9ccee153a0ae" \
-go test ./...
+go test -v ./...
