+++ tests/director/tools/example-index-generator/README.md
