+++ tests/director/tools/example-index-generator/main.go
