+++ tests/director/tools/example-index-generator/md.tpl
