+++ tests/director/wait-for-director.sh
